requirements.txt
# 17lands Data Project
duckdb>=0.9.0
pyyaml>=6.0
requests>=2.28.0

# dbt (install separately: pip install dbt-duckdb)
# dbt-duckdb>=1.8.0
